Turn MojoShell into an actual binding
