+++ packages/flutter/lib/src/widgets/binding.dart
@@ -21,7 +21,7 @@ class BindingObserver {
 
 /// A concrete binding for applications based on the Widgets framework.
 /// This is the glue that binds the framework to the Flutter engine.
-class WidgetFlutterBinding extends BindingBase with Scheduler, Gesturer, Renderer {
+class WidgetFlutterBinding extends BindingBase with Scheduler, Gesturer, MojoShell, Renderer {
 
   WidgetFlutterBinding._();
